% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataseries-package.R
\docType{package}
\name{dataseries-package}
\alias{dataseries}
\alias{dataseries-package}
\title{dataseries: Switzerland's Data Series in One Place}
\description{
Download and import open Swiss economic time series from
\href{https://dataseries.org}{dataseries.org}.
}
\details{
The data are organized as \strong{datasets}, each of which is a family of related
series. Most datasets are multi-dimensional \emph{cubes}: a single time series is
one cell, addressed by the dataset plus one code per dimension (e.g. the GDP
cube splits along \code{type}, \code{structure} and \code{seas_adj}). A handful of datasets
are a single series and need no dimensions at all.
\itemize{
\item \code{\link[=ds_catalog]{ds_catalog()}} lists every available dataset.
\item \code{\link[=ds_search]{ds_search()}} is a flat, searchable list of the individual series.
\item \code{\link[=ds_meta]{ds_meta()}} describes one dataset's dimensions and their codes.
\item \code{\link[=ds]{ds()}} downloads series, as a \code{data.frame} or \code{ts} object.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dataseries.org}
  \item \url{https://github.com/christophsax/dataseries}
  \item Report bugs at \url{https://github.com/christophsax/dataseries/issues}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph.sax@gmail.com}

Other contributors:
\itemize{
  \item Jannes Muenchow \email{malnamalja@gmx.de} [contributor]
}

}
\keyword{internal}
