% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{Download time series from
\href{http://www.dataseries.org}{www.dataseries.org}}
\usage{
ds(id, class = c("data.frame", "xts"))
}
\arguments{
\item{id}{one ore more IDs, as shown on
\href{http://www.dataseries.org}{www.dataseries.org}.}

\item{class}{class of the return value, either a \code{"data.frame"} (default) or
an \code{"xts"} object.}
}
\description{
Download time series from
\href{http://www.dataseries.org}{www.dataseries.org} as a \code{data.frame} or and
\code{xts} object.
}
\details{
Downloaded series are \strong{cached in the memory} and kept there as long as your
R session is open. If you want to force an update, either restart your R
session or delete the caching enivronment, \code{.in.memory.cache} (see examples).
}
\examples{
/dontrun{
ds(c("CCI.AIK", "CCI.ASSS"))
ds(c("CCI.AIK", "CCI.ASSS"), class = "xts")

# delete in-memory cache, to force a fresh download
rm(".in.memory.cache")
}

}
