% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory.R
\name{inventory}
\alias{inventory}
\title{Inventory from \href{http://www.dataseries.org}{www.dataseries.org}}
\usage{
inventory()
}
\description{
Read latest time series inventory from
\href{http://www.dataseries.org}{www.dataseries.org} as a \code{data.frame}.
}
\details{
The downloaded data are \strong{cached in memory} as long as the
R session is open. If you want to force an update, either restart the R
session or run \code{cache_rm} to emtpy the cache. \code{cache_ls} is a utility
function that lists all cached objects.
}
\examples{
\dontrun{
inv <- inventory()  # will be cached
# search in the inventory
inv[grepl("consumer conf", inv$description, ignore.case = TRUE), ]
}

}
